+++ doc/user_manual/extractCommentsFromRbelValidatorGlue.sh
@@ -1,0 +1,9 @@
+#
+# ${GEMATIK_COPYRIGHT_STATEMENT}
+#
+
+sed 's/^ *//;s/ *$//' < ../../tiger-test-lib/src/test/java/de/gematik/test/tiger/glue/RBelValidatorGlue.java | \
+grep -e "^[\*|\/|\@]" | \
+grep -ve "^//" | \
+sed 's/^\/\*/\n\//' | \
+tail -n +6 > RbelValidatorGlueCommentsOnly.java